refactor: extract admin durable audit store
