+++ crates/dcc-mcp-gateway-admin/Cargo.toml
@@ -31,3 +31,4 @@ dunce = "1.0"
 [dev-dependencies]
 dcc-mcp-test-utils.workspace = true
 proptest = "1.5"
+tempfile.workspace = true
