+++ crates/dcc-mcp-gateway-admin/src/lib.rs
@@ -2,7 +2,8 @@
 //!
 //! This crate owns admin-facing audit, trace, caller-context, compact projection,
 //! link, issue-report, statistics, analytics, governance, artifact, activity, task-outcome,
-//! debug-bundle, postmortem, agent-trace packet, memory-summary, experiment, skill-path, and traffic projections
+//! durable audit, debug-bundle, postmortem, agent-trace packet, memory-summary, experiment,
+//! skill-path, and traffic projections
 //! independently of gateway routing state.
 //! It also owns the Vite/npm build script and generated dashboard payload; the Node.js
 //! toolchain only runs when `embed` is enabled.
@@ -17,6 +18,7 @@ mod audit;
 mod debug;
 /// Admin trace and caller-attribution value types.
 pub mod domain;
+mod durable_store;
 mod experiments;
 mod governance;
 mod issue_report;
@@ -54,6 +56,7 @@ pub use domain::trace::{
     TokenTelemetry, TraceContext, TraceContextHeader, TracePayload, TraceSpan, estimate_tokens,
     parse_traceparent,
 };
+pub use durable_store::DurableAuditStore;
 pub use experiments::{
     ExperimentJudgeValidation, project_experiment_detail, project_experiment_list,
     valid_experiment_id, validate_experiment_definition, validate_experiment_judge,
